filtering at the line level works but the recognizer is imperfect, generic top level application of PresidioFilter is not working as expected. The spacy library is not properly linked ot presidio and downloads every time the frontend restarts. likely a uv issue
